Add ignored plugin support to auto update script
This script updates the plugins via the FriendsOfGalaxy update system.
This system has been broken for several months and no plugins are
receiving updates.

Any of these "official" plugins with upstream updates have been updated
in the main manifest yaml file to point to the upstream repository
instead. Therefore, we need to ignore these plugins as part of the auto
update script so they aren't downgraded to the FriendsOfGalaxy version,
which is out of date.